+++ test/libsolidity/syntaxTests/parsing/import_empty.sol
@@ -1,0 +1,3 @@
+import "";
+// ----
+// ParserError: (9-10): Import path cannot be empty.